+++ ksa_compliance/standard_doctypes/sales_invoice.py
@@ -1,8 +1,19 @@
 import frappe
+
+from ksa_compliance import logger
 from ksa_compliance.ksa_compliance.doctype.zatca_business_settings.zatca_business_settings import ZATCABusinessSettings
 
 
 def create_sales_invoice_additional_fields_doctype(self, method):
+    settings = ZATCABusinessSettings.for_invoice(self.name)
+    if not settings:
+        logger.info(f"Skipping additional fields for sales invoice {self.name} because of missing ZATCA settings")
+        return
+
+    if not settings.has_production_csid:
+        logger.info(f"Skipping additional fields for sales invoice {self.name} because of missing production CSID")
+        return
+
     si_additional_fields_doc = frappe.new_doc("Sales Invoice Additional Fields")
     si_additional_fields_doc.sales_invoice = self.name
     if self.customer_address:
@@ -20,7 +31,6 @@ def create_sales_invoice_additional_fields_doctype(self, method):
         si_additional_fields_doc.update(address_info)
 
     # We have to insert before submitting to ensure we can properly update the document with the hash, XML, etc.
-    settings = ZATCABusinessSettings.for_invoice(self.name)
     if settings.is_live_sync:
         si_additional_fields_doc.insert()
         si_additional_fields_doc.submit()